requirements.txt
alive_progress==1.6.1
matplotlib==3.3.2
numpy==1.19.2
Pillow==7.2.0
PyQt5==5.15.1
pytesseract==0.3.6
selenium==3.141.0
